% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density function for the Triangular Distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{min}{Lower limit of the distribution (a).}

\item{max}{Upper limit of the distribution (b).}

\item{mode}{Mode of the distribution (c).}
}
\value{
A numeric vector of densities.
}
\description{
Calculates the density at x for a triangular distribution with
parameters min, max, and mode.
}
\examples{
# Calcular la densidad en x = 0.5 para un triángulo entre 0 y 1, con pico en 0.5
dtriang(x = 0.5, a = 0, b = 1, c = 0.5)
}
