% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function for the Triangular Distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{min}{Lower limit (a).}

\item{max}{Upper limit (b).}

\item{mode}{Mode (c).}
}
\value{
A numeric vector of cumulative probabilities.
}
\description{
Cumulative distribution function (CDF).
}
\examples{
# Probabilidad acumulada hasta x = 0.5
ptriang(q = 0.5, a = 0, b = 1, c = 0.5)
}
