% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function for the Triangular Distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{min}{Lower limit (a).}

\item{max}{Upper limit (b).}

\item{mode}{Mode (c).}
}
\value{
A numeric vector of quantiles.
}
\description{
Returns the quantile for a given probability p.
}
\examples{
# Encontrar el valor de x que acumula el 50\% de la probabilidad (mediana)
qtriang(p = 0.5, a = 0, b = 1, c = 0.5)
}
