% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation for the Triangular Distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{min}{Lower limit of the distribution (a).}

\item{max}{Upper limit of the distribution (b).}

\item{mode}{Mode of the distribution (c).}
}
\value{
A numeric vector of random values.
}
\description{
Generates n random values from a triangular distribution using the inversion method.
}
\examples{
# Generar 10 valores aleatorios de esta distribución triangular
rtriang(n = 10, a = 0, b = 1, c = 0.5)
}
